+++ src/app/Home/page.tsx
@@ -105,9 +105,6 @@ export default function Home() {
                 </motion.div>
 
             </motion.div>
-
-            {/* ? Footer */}
-            <Footer />
         </div>
     );
 }